Address review: extract ApplicationDevelopmentService and shared app token fetcher
Moves the application development mutation logic out of the resolver into a
dedicated service, and deduplicates the generateApplicationToken fetchTokenPair
callback between dev-once and the dev-mode orchestrator.
