+++ packages/twenty-server/src/engine/core-modules/application/application-development/application-development.module.ts
@@ -5,6 +5,7 @@ import { ApplicationManifestModule } from 'src/engine/core-modules/application/a
 import { ApplicationModule } from 'src/engine/core-modules/application/application.module';
 import { ApplicationPackageModule } from 'src/engine/core-modules/application/application-package/application-package.module';
 import { ApplicationDevelopmentResolver } from 'src/engine/core-modules/application/application-development/application-development.resolver';
+import { ApplicationDevelopmentService } from 'src/engine/core-modules/application/application-development/application-development.service';
 import { CacheLockModule } from 'src/engine/core-modules/cache-lock/cache-lock.module';
 import { FeatureFlagModule } from 'src/engine/core-modules/feature-flag/feature-flag.module';
 import { FileStorageModule } from 'src/engine/core-modules/file-storage/file-storage.module';
@@ -28,6 +29,7 @@ import { WorkspaceMigrationGraphqlApiExceptionInterceptor } from 'src/engine/wor
   ],
   providers: [
     ApplicationDevelopmentResolver,
+    ApplicationDevelopmentService,
     WorkspaceMigrationGraphqlApiExceptionInterceptor,
   ],
 })
